Refactoring setup project to build two docker images (basex and postgresql). Flyway is now configured using Quarkus.
