+++ app/src/main/docker/Dockerfile-basex.jvm
@@ -44,10 +44,10 @@ RUN microdnf install curl ca-certificates ${JAVA_PACKAGE} \
 ENV JAVA_OPTIONS="-Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager"
 ENV JAVA_MAX_MEM_RATIO="70"
 # We make four distinct layers so if there are application changes the library layers can be re-used
-COPY --chown=1001 target/quarkus-app/lib/ /deployments/lib/
-COPY --chown=1001 target/quarkus-app/*.jar /deployments/
-COPY --chown=1001 target/quarkus-app/app/ /deployments/app/
-COPY --chown=1001 target/quarkus-app/quarkus/ /deployments/quarkus/
+COPY --chown=1001 target/basex-quarkus-app/lib/ /deployments/lib/
+COPY --chown=1001 target/basex-quarkus-app/*.jar /deployments/
+COPY --chown=1001 target/basex-quarkus-app/app/ /deployments/app/
+COPY --chown=1001 target/basex-quarkus-app/quarkus/ /deployments/quarkus/
 
 EXPOSE 8080
 USER 1001
