+++ app/src/main/docker/Dockerfile-postgresql.jvm
@@ -3,22 +3,22 @@
 #
 # Before building the container image run:
 #
-# ./mvnw package -Dquarkus.package.type=legacy-jar
+# ./mvnw package
 #
 # Then, build the image with:
 #
-# docker build -f src/main/docker/Dockerfile.legacy-jar -t quarkus/app-legacy-jar .
+# docker build -f src/main/docker/Dockerfile.jvm -t quarkus/app-jvm .
 #
 # Then run the container using:
 #
-# docker run -i --rm -p 8080:8080 quarkus/app-legacy-jar
+# docker run -i --rm -p 8080:8080 quarkus/app-jvm
 #
 # If you want to include the debug port into your docker image
 # you will have to expose the debug port (default 5005) like this :  EXPOSE 8080 5005
 #
 # Then run the container using :
 #
-# docker run -i --rm -p 8080:8080 -p 5005:5005 -e JAVA_ENABLE_DEBUG="true" quarkus/app-legacy-jar
+# docker run -i --rm -p 8080:8080 -p 5005:5005 -e JAVA_ENABLE_DEBUG="true" quarkus/app-jvm
 #
 ###
 FROM registry.access.redhat.com/ubi8/ubi-minimal:8.4 
@@ -43,10 +43,14 @@ RUN microdnf install curl ca-certificates ${JAVA_PACKAGE} \
 # Configure the JAVA_OPTIONS, you can add -XshowSettings:vm to also display the heap size.
 ENV JAVA_OPTIONS="-Dquarkus.http.host=0.0.0.0 -Djava.util.logging.manager=org.jboss.logmanager.LogManager"
 ENV JAVA_MAX_MEM_RATIO="70"
-COPY target/lib/* /deployments/lib/
-COPY target/*-runner.jar /deployments/app.jar
+# We make four distinct layers so if there are application changes the library layers can be re-used
+COPY --chown=1001 target/postgresql-quarkus-app/lib/ /deployments/lib/
+COPY --chown=1001 target/postgresql-quarkus-app/*.jar /deployments/
+COPY --chown=1001 target/postgresql-quarkus-app/app/ /deployments/app/
+COPY --chown=1001 target/postgresql-quarkus-app/quarkus/ /deployments/quarkus/
 
 EXPOSE 8080
 USER 1001
 
 ENTRYPOINT [ "/deployments/run-java.sh" ]
+
