+++ app/src/main/docker/Dockerfile-postgresql.native
@@ -19,7 +19,7 @@ WORKDIR /work/
 RUN chown 1001 /work \
     && chmod "g+rwX" /work \
     && chown 1001:root /work
-COPY --chown=1001:root target/*-runner /work/application
+COPY --chown=1001:root target/postgresql-quarkus-app/*-runner /work/application
 
 EXPOSE 8080
 USER 1001
