+++ app/src/main/resources/application-dev-basex.properties
@@ -1,0 +1,28 @@
+# SPDX-FileCopyrightText: 2021 Alliander N.V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# Development BaseX configuration.
+quarkus.http.port = 9090
+quarkus.http.cors = true
+
+# Dev Logging
+quarkus.log.level = DEBUG
+quarkus.log.category."org.lfenergy.compas.scl.data".level = DEBUG
+
+# Dev extra permissions
+quarkus.http.auth.permission.develop-quarkus-services.paths  = /compas-scl-data-service/q/swagger-ui/*,/compas-scl-data-service/index.html
+quarkus.http.auth.permission.develop-quarkus-services.policy = permit
+
+# Datasource configuration for BaseX (none)
+quarkus.datasource.jdbc                = false
+quarkus.datasource.devservices.enabled = false
+
+# Flyway configuration for BaseX (none)
+quarkus.flyway.migrate-at-start = false
+
+# BaseX configuration
+basex.host              = localhost
+basex.port              = 1984
+basex.username          = admin
+basex.password          = admin
