+++ app/src/main/resources/application-dev-postgresql.properties
@@ -1,0 +1,27 @@
+# SPDX-FileCopyrightText: 2021 Alliander N.V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# Development PostgreSQL configuration.
+quarkus.http.port = 9090
+quarkus.http.cors = true
+
+# Dev Logging
+quarkus.log.level = DEBUG
+quarkus.log.category."org.lfenergy.compas.scl.data".level = DEBUG
+
+# Dev extra permissions
+quarkus.http.auth.permission.develop-quarkus-services.paths  = /compas-scl-data-service/q/swagger-ui/*,/compas-scl-data-service/index.html
+quarkus.http.auth.permission.develop-quarkus-services.policy = permit
+
+# Datasource configuration for PostgreSQL
+quarkus.datasource.devservices.enabled = false
+quarkus.datasource.db-kind             = postgresql
+quarkus.datasource.jdbc.url            = jdbc:postgresql://localhost:5432/compas
+quarkus.datasource.jdbc.max-size       = 16
+quarkus.datasource.username            = postgres
+quarkus.datasource.password            = postgres
+
+# Flyway configuration for PostgreSQL
+quarkus.flyway.migrate-at-start = true
+quarkus.flyway.locations        = classpath:org/lfenergy/compas/scl/data/repository/postgresql/db/migration
