+++ app/src/main/resources/application-prod-basex.properties
@@ -1,0 +1,21 @@
+# SPDX-FileCopyrightText: 2021 Alliander N.V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# Production BaseX configuration.
+
+# Add scanning these dependencies for scanning, also used by native compilation.
+quarkus.index-dependency.scl-data-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.index-dependency.scl-data-repository.artifact-id = repository-basex
+
+# Datasource configuration for BaseX (none)
+quarkus.datasource.jdbc = false
+
+# Flyway configuration for BaseX (none)
+quarkus.flyway.migrate-at-start = false
+
+# BaseX configuration
+basex.host              = localhost
+basex.port              = 1984
+basex.username          = admin
+basex.password          = admin
