+++ app/src/main/resources/application-prod-postgresql.properties
@@ -1,0 +1,20 @@
+# SPDX-FileCopyrightText: 2021 Alliander N.V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# Production PostgreSQL configuration.
+
+# Add scanning these dependencies for scanning, also used by native compilation.
+quarkus.index-dependency.scl-data-repository.group-id    = org.lfenergy.compas.scl.data
+quarkus.index-dependency.scl-data-repository.artifact-id = repository-postgresql
+
+# Datasource configuration for PostgreSQL
+quarkus.datasource.db-kind       = postgresql
+quarkus.datasource.jdbc.url      = jdbc:postgresql://localhost:5432/compas
+quarkus.datasource.jdbc.max-size = 16
+quarkus.datasource.username      = postgres
+quarkus.datasource.password      = postgres
+
+# Flyway configuration for PostgreSQL
+quarkus.flyway.migrate-at-start = true
+quarkus.flyway.locations        = classpath:org/lfenergy/compas/scl/data/repository/postgresql/db/migration
