+++ app/src/main/resources/application-test.properties
@@ -1,0 +1,13 @@
+# SPDX-FileCopyrightText: 2021 Alliander N.V.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+# Test common configuration.
+
+# Datasource configuration for BaseX (none)
+quarkus.datasource.jdbc                = false
+quarkus.datasource.devservices.enabled = false
+
+# Flyway configuration for BaseX (none)
+quarkus.flyway.migrate-at-start = false
+
