+++ app/src/main/resources/application.properties
@@ -14,26 +14,15 @@ quarkus.http.limits.max-body-size = 150M
 quarkus.log.level = INFO
 quarkus.log.category."org.lfenergy.compas.scl.data".level = INFO
 
-# Dev Profile overrides.
-%dev.quarkus.http.port = 9090
-%dev.quarkus.http.cors = true
+# Add scanning these dependencies for scanning, also used by native compilation.
+quarkus.index-dependency.jaxrs-commons.group-id=org.lfenergy.compas.core
+quarkus.index-dependency.jaxrs-commons.artifact-id=jaxrs-commons
 
-%dev.quarkus.log.level = DEBUG
-%dev.quarkus.log.category."org.lfenergy.compas.scl.data".level = DEBUG
+quarkus.index-dependency.scl-data-service.group-id=org.lfenergy.compas.scl.data
+quarkus.index-dependency.scl-data-service.artifact-id=service
 
-# BaseX configuration
-%prod-basex.basex.host      = ${BASEX_HOST:localhost}
-%prod-basex.basex.port      = ${BASEX_PORT:1984}
-%prod-basex.basex.username  = ${BASEX_USERNAME:admin}
-%prod-basex.basex.password  = ${BASEX_PASSWORD:admin}
-
-# Datasource configuration for PostgreSQL
-%prod-postgres.quarkus.datasource.db-kind       = others
-%prod-postgres.quarkus.datasource.jdbc.driver   = org.postgresql.Driver
-%prod-postgres.quarkus.datasource.jdbc.url      = jdbc:postgresql://localhost:5432/compas
-%prod-postgres.quarkus.datasource.jdbc.max-size = 16
-%prod-postgres.quarkus.datasource.username      = postgres
-%prod-postgres.quarkus.datasource.password      = postgres
+quarkus.index-dependency.scl-data-repository.group-id=org.lfenergy.compas.scl.data
+quarkus.index-dependency.scl-data-repository.artifact-id=repository
 
 # Smallrye JWT Properties (Microprofile)
 smallrye.jwt.verify.key.location = ${JWT_VERIFY_KEY:http://localhost:8089/auth/realms/compas/protocol/openid-connect/certs}
@@ -46,9 +35,6 @@ quarkus.http.auth.permission.deny-default.policy=deny
 
 quarkus.http.auth.permission.allow-quarkus-services.paths=/compas-scl-data-service/q/health/live,/compas-scl-data-service/q/health/ready,/compas-scl-data-service/q/openapi
 quarkus.http.auth.permission.allow-quarkus-services.policy=permit
-
-%dev.quarkus.http.auth.permission.develop-quarkus-services.paths=/compas-scl-data-service/q/swagger-ui/*,/compas-scl-data-service/index.html
-%dev.quarkus.http.auth.permission.develop-quarkus-services.policy=permit
 
 quarkus.http.auth.permission.common.paths=/compas-scl-data-service/common/v1/*
 quarkus.http.auth.permission.common.policy=authenticated
