+++ pom.xml
@@ -31,7 +31,8 @@ SPDX-License-Identifier: Apache-2.0
         <microprofile-config-api.version>2.0</microprofile-config-api.version>
         <microprofile-openapi-api.version>3.0-RC5</microprofile-openapi-api.version>
         <apache.commons-io.version>2.11.0</apache.commons-io.version>
-        <flyway.version>8.0.4</flyway.version>
+        <flyway.version>7.15.0</flyway.version>
+        <postgresql.version>42.3.1</postgresql.version>
         <slf4j.version>1.7.32</slf4j.version>
 
         <!-- Test -->
@@ -135,6 +136,11 @@ SPDX-License-Identifier: Apache-2.0
                 <groupId>org.flywaydb</groupId>
                 <artifactId>flyway-core</artifactId>
                 <version>${flyway.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>org.postgresql</groupId>
+                <artifactId>postgresql</artifactId>
+                <version>${postgresql.version}</version>
             </dependency>
 
             <dependency>
