+++ repository-basex/src/main/java/org/lfenergy/compas/scl/data/basex/client/BaseXClientFactory.java
@@ -3,24 +3,18 @@
 // SPDX-License-Identifier: Apache-2.0
 package org.lfenergy.compas.scl.data.basex.client;
 
-import org.eclipse.microprofile.config.inject.ConfigProperty;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Inject;
 import java.io.IOException;
 
-@ApplicationScoped
 public class BaseXClientFactory {
     private final String baseXHost;
     private final Integer baseXPort;
     private final String baseXUsername;
     private final String baseXPassword;
 
-    @Inject
-    public BaseXClientFactory(@ConfigProperty(name = "basex.host") String baseXHost,
-                              @ConfigProperty(name = "basex.port") Integer baseXPort,
-                              @ConfigProperty(name = "basex.username") String baseXUsername,
-                              @ConfigProperty(name = "basex.password") String baseXPassword) {
+    public BaseXClientFactory(String baseXHost,
+                              Integer baseXPort,
+                              String baseXUsername,
+                              String baseXPassword) {
         this.baseXHost = baseXHost;
         this.baseXPort = baseXPort;
         this.baseXUsername = baseXUsername;
